+++ mono_repo/lib/src/commands/travis/shared.dart
@@ -16,3 +16,4 @@ function $commandName() {
     command $commandName "\$@"
   fi
 }''';
+